backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
python-dotenv>=1.0.1
numpy>=1.24.0
h5py>=3.9.0
nltk>=3.8.1
scipy>=1.10.0
loguru>=0.7.3
requests>=2.32.0
moviepy==2.1.1
torch>=2.1.0
edge-tts==7.2.7
streamlit>=1.45.0
pydub==0.25.1
pysrt==1.1.2
openai>=1.77.0
google-generativeai>=0.8.5
azure-cognitiveservices-speech>=1.37.0
tencentcloud-sdk-python>=3.0.1200
dashscope>=1.24.6
Pillow>=10.3.0
tqdm>=4.66.6
tenacity>=9.0.0
watchdog==6.0.0
toml>=0.10.2
tomli>=2.2.1
tomli-w>=1.0.0
pyJianYingDraft>=0.1.0
